App & client: set tls.Config.ServerName for every connection rather than once
Because of https://github.com/FiloSottile/go/commit/268080 ServerName
needs to be set in the TLS config in order to Write() directly to a TLS
connection. This chance tasks dialers to set the proper ServerName when
making a connection rather than setting it once for every connection

This lets each dqlite node use its own certificate when running in a cluster.
The certificates are changed to have their SAN IP set to 127.0.0.1 as this is
the address used in unit tests