Update go-sqlite3 package to v1.14.5
Somewhat counterintuitively the latest version of go-sqlite3 is v1.14.5
as opposed to v2.0.3. Apparently updating to v2 was a mistake, as noted
in https://github.com/mattn/go-sqlite3/blob/master/README.md#go-sqlite3.
Trying to use the latest versions of go-dqlite and go-sqlite3 in a
project resulted in a kind of go.mod edit war. This therefore updates
the go-sqlite3 dependency to the latest version.