Provide tracing options at the app level
The following provides the tracing options at the app level, so we can
then pass them directly to the driver. This will allow monitoring of the
traces provided by the driver.

This approach will suffice for now, but might be better to rethink how
we expose driver capabilities in the future.

Fixes: #242

Signed-off-by: Simon Richardson <stickupkid@gmail.com>