Allow BEGIN CONCURRENT for transactions
The following implements a TxMode for testing purposes. The idea
being we would like to see if there is any performance difference
in using BEGIN CONCURRENT or not in juju.

The notion is simple, we pass a option through the App and to the
driver and based on the value of the option we switch the tx start
mode.

Signed-off-by: Simon Richardson <stickupkid@gmail.com>