Always use QUERY_SQL request in the dqlite shell
In recent versions of dqlite, the QUERY_SQL request supports SQL
queries that modify the database and do not return rows. Update
internal/shell/shell.go to always send this request for SQL strings
entered by the user, instead of switching between QUERY_SQL and EXEC_SQL
based on whether the string start with `SELECT`.

Also fix the Rows iterator so that it can deal with empty row sets that
have zero columns. Previously it would enter an infinite loop in this
situation.

Signed-off-by: Cole Miller <cole.miller@canonical.com>