OpenTelemetry Driver Tracing
To aid with debugging and development, the following implements the
ability to trace requests to the driver. This includes adding the
current SQL from conn methods ExecContext or QueryContext, and
also adding the stmt ExecContext and QueryContext. Unfortunately,
we can't trace every method as the go driver library doesn't allow
passing a context to Commit or Rollback.

The tracing code does not require any external dependencies, as this
retrieves a Tracer from the passed in context. As long as the Tracer
conforms to the tracing.Tracer interface then a trace can be performed.

There will always be a Tracer if one isn't located on the context.

Potentially, this could supersede the current tracing implementation
and remove the need to have a tracing log.