Remove TracerWithContext
There should be no need to expose the TracerWithContext from dqlite
directly. Instead the tracer should only be added to the context and
internally, we can just use the key. This removes the need to use
an intermidate noop tracer entity, we can go directly to the noop
span.

This reduces the code surface even more, which is a bonus for
maintainability.