internal/protocol: Check if leader was found in between network requests
We send a request to multiple peers concurrently to ask if they are the
leader. They themselves may be the leader, or they may report who they
think is the leader. We then try to reach out to whoever they think is
the leader.

It is possible that we are are already talking to the leader
concurrently with asking a different node who the leader is, so in such
cases we don't need to send further network requests and can return
early by checking if we have already found the leader in another
goroutine.

Signed-off-by: Max Asnaashari <max.asnaashari@canonical.com>