feat: tab writer
Without column headers it's non-obvious what each column value is
without referring to the original schema. To remove the ambiguity
and make it easier to read the output this adds a tabwriter from
the standard library.

The output is a lot cleaner IMO, but you do loose the separator
lines. We could add it back using an external package, but I'm
unsure it's worth it.

Fixes: #33