Fixes
First, avoid infinite recursion when logging in connectAttemptOne.

Second, fix the tests by logging the error from Semaphore.Acquire. It
seems that before we bumped the version of golang.org/x/sync/semaphore,
this type didn't honor the context deadline, hence why the tests were
fine for me locally without merging in the go.mod updates that are on
master.

Signed-off-by: Cole Miller <cole.miller@canonical.com>